# perf: turn off jsx transforms for non-jsx compatible files