@dsherret 2024-05-30
Done here: https://github.com/denoland/deno_ast/pull/252/commits/c9d89dccddaef38000869690247b258b779d2bf2